renaming dir and fixing 080 index
